refactor: separate the TrackingIdToLatestHash method to a separate file
